refactor: move existing provider to v1 folder
